refactor(payments): Move processNonZeroInvoice, processZeroInvoice, and processInvoice to InvoiceManager
