+++ libs/payments/paypal/src/lib/paypal.manager.ts
@@ -6,10 +6,6 @@ import { Injectable } from '@nestjs/common';
 
 import {
   ACTIVE_SUBSCRIPTION_STATUSES,
-  CustomerManager,
-  InvoiceManager,
-  StripeCustomer,
-  StripeInvoice,
   SubscriptionManager,
 } from '@fxa/payments/stripe';
 import { PayPalClient } from './paypal.client';
@@ -23,9 +19,7 @@ import { PaypalManagerError } from './paypal.error';
 export class PayPalManager {
   constructor(
     private client: PayPalClient,
-    private customerManager: CustomerManager,
     private subscriptionManager: SubscriptionManager,
-    private invoiceManager: InvoiceManager,
     private paypalCustomerManager: PaypalCustomerManager
   ) {}
 
@@ -137,50 +131,6 @@ export class PayPalManager {
         ACTIVE_SUBSCRIPTION_STATUSES.includes(sub.status) &&
         sub.collection_method === 'send_invoice'
     );
-  }
-
-  /**
-   * Process an invoice when amount is greater than minimum amount
-   */
-  async processNonZeroInvoice(
-    customer: StripeCustomer,
-    invoice: StripeInvoice,
-    ipaddress?: string
-  ) {
-    // TODO in M3b: Implement legacy processInvoice as processNonZeroInvoice here
-    // TODO: Add spec
-    console.log(customer, invoice, ipaddress);
-  }
-
-  /**
-   * Finalize and process a draft invoice that has no amounted owed.
-   */
-  async processZeroInvoice(invoiceId: string) {
-    // It appears for subscriptions that do not require payment, the invoice
-    // transitions to paid automatially.
-    // https://stripe.com/docs/billing/invoices/subscription#sub-invoice-lifecycle
-    return this.invoiceManager.finalizeWithoutAutoAdvance(invoiceId);
-  }
-
-  /**
-   * Process an invoice
-   * If amount is less than minimum amount, call processZeroInvoice
-   * If amount is greater than minimum amount, call processNonZeroInvoice (legacy PaypalHelper processInvoice)
-   */
-  async processInvoice(invoice: StripeInvoice) {
-    if (!invoice.customer) throw new Error('Customer not present on invoice');
-    const amountInCents = invoice.amount_due;
-
-    if (
-      amountInCents <
-      this.subscriptionManager.getMinimumAmount(invoice.currency)
-    ) {
-      await this.processZeroInvoice(invoice.id);
-      return;
-    }
-
-    const customer = await this.customerManager.retrieve(invoice.customer);
-    await this.processNonZeroInvoice(customer, invoice);
   }
 
   /*
